refactor: replace while loop with for

+++ backend/src/main/java/it/compare/backend/scraping/rtveuroagd/dto/RtvEuroAgdResponse.java
@@ -2,4 +2,4 @@ package it.compare.backend.scraping.rtveuroagd.dto;
 
 import java.util.List;
 
-public record RtvEuroAgdResponse(Long productsCount, List<RtvEuroAgdProduct> results) {}
+public record RtvEuroAgdResponse(Integer productsCount, List<RtvEuroAgdProduct> results) {}
